requirements.txt
appdirs==1.4.4
asgiref==3.2.10
astroid==2.4.1
beautifulsoup4==4.9.1
bs4==0.0.1
click==7.1.2
distlib==0.3.1
Django==3.1
django-filter==2.3.0
django-tables2==2.3.1
filelock==3.0.12
Flask==1.1.2
Flask-Session==0.3.1
Flask-SocketIO==4.3.0
Flask-SQLAlchemy==2.4.1
html5lib==1.1
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
lxml==4.5.2
MarkupSafe==1.1.1
mccabe==0.6.1
psycopg2-binary==2.8.5
pylint==2.5.2
python-engineio==3.12.1
python-socketio==4.5.1
pytz==2020.1
six==1.14.0
soupsieve==2.0.1
SQLAlchemy==1.3.16
sqlparse==0.3.1
toml==0.10.0
virtualenv==20.0.30
webencodings==0.5.1
Werkzeug==0.16.0
wrapt==1.12.1
